+++ tests/test_penman.py
@@ -4,12 +4,12 @@ import penman
 def test_from_penman():
     parse = penman.Graph.from_penman
     g = parse('(h / hi)')
-    assert g.to_triples() == [('h', 'instance-of', 'hi')]
+    assert g.triples() == [('h', 'instance', 'hi')]
     g = parse('(b / bark :ARG0 (d / dog))')
-    assert sorted(g.to_triples()) == [
+    assert g.triples() == [
+        ('b', 'instance', 'bark'),
         ('b', 'ARG0', 'd'),
-        ('b', 'instance-of', 'bark'),
-        ('d', 'instance-of', 'dog')
+        ('d', 'instance', 'dog')
     ]
     g = parse(
         '(e2 / _try_v_1'
@@ -18,48 +18,45 @@ def test_from_penman():
         '              :RSTR-of (_1 / proper_q))'
         '    :ARG2 (e3 / _sleep_v_1'
         '              :ARG1 x1))')
-    assert sorted(g.to_triples()) == [
-        ('_1', 'instance-of', 'proper_q'),
+    assert g.triples() == [
+        ('e2', 'instance', '_try_v_1'),
         ('e2', 'ARG1', 'x1'),
-        ('e2', 'ARG2', 'e3'),
-        ('e2', 'instance-of', '_try_v_1'),
-        ('e3', 'ARG1', 'x1'),
-        ('e3', 'instance-of', '_sleep_v_1'),
+        ('x1', 'instance', 'named'),
         ('x1', 'CARG', '"Abrams"'),
-        ('x1', 'RSTR-of', '_1'),
-        ('x1', 'instance-of', 'named')
+        ('_1', 'RSTR', 'x1'),
+        ('_1', 'instance', 'proper_q'),
+        ('e2', 'ARG2', 'e3'),
+        ('e3', 'instance', '_sleep_v_1'),
+        ('e3', 'ARG1', 'x1'),
     ]
+    
 
 def test_to_penman():
     load = penman.Graph.from_triples
     g = load([
-        ('h', 'instance-of', 'hi')
+        ('h', 'instance', 'hi')
     ])
     assert g.to_penman() == '(h / hi)'
-    g = load([
-        ('hi', 'instance', 'h')
-    ])
-    assert g.to_penman() == '(hi :instance h)'
     assert g.to_penman(top='h') == '(h / hi)'
     g = load([
         ('b', 'ARG0', 'd'),
-        ('b', 'instance-of', 'bark'),
-        ('d', 'instance-of', 'dog')
+        ('b', 'instance', 'bark'),
+        ('d', 'instance', 'dog')
     ])
     assert g.to_penman() == (
         '(b / bark\n'
         '   :ARG0 (d / dog))'
     )
     g = load([
-        ('e2', 'instance-of', '_try_v_1'),
+        ('e2', 'instance', '_try_v_1'),
         ('e2', 'ARG1', 'x1'),
         ('e2', 'ARG2', 'e3'),
-        ('x1', 'instance-of', 'named'),
+        ('x1', 'instance', 'named'),
         ('x1', 'CARG', '"Abrams"'),
-        ('x1', 'RSTR-of', '_1'),
-        ('e3', 'instance-of', '_sleep_v_1'),
+        ('_1', 'RSTR', 'x1'),
+        ('e3', 'instance', '_sleep_v_1'),
         ('e3', 'ARG1', 'x1'),
-        ('_1', 'instance-of', 'proper_q')
+        ('_1', 'instance', 'proper_q')
     ])
     assert g.to_penman() == (
         '(e2 / _try_v_1\n'
@@ -70,13 +67,13 @@ def test_to_penman():
         '              :ARG1 x1))'
     )
     g = load([
-        ('a', 'instance-of', 'aaa'),
+        ('a', 'instance', 'aaa'),
         ('a', 'ARG1', 'b'),
-        ('b', 'instance-of', 'bbb'),
-        ('c', 'instance-of', 'ccc'),
+        ('b', 'instance', 'bbb'),
+        ('c', 'instance', 'ccc'),
         ('c', 'ARG1', 'b2'),
         ('c', 'ARG2', 'a'),
-        ('b2', 'instance-of', 'bbb')
+        ('b2', 'instance', 'bbb')
     ])
     assert g.to_penman() == (
         '(a / aaa\n'
@@ -87,11 +84,11 @@ def test_to_penman():
 
 def test_to_penman_with_indent():
     g = penman.Graph.from_triples([
-        ('a', 'instance-of', 'aaa'),
+        ('a', 'instance', 'aaa'),
         ('a', 'ARG1', 'b'),
-        ('b', 'instance-of', 'bbb'),
+        ('b', 'instance', 'bbb'),
         ('b', 'ARG1', 'c'),
-        ('c', 'instance-of', 'ccc')
+        ('c', 'instance', 'ccc')
     ])
     assert g.to_penman(indent=True) == (
         '(a / aaa\n'
@@ -145,7 +142,6 @@ def test_graph_inspection():
         ('e3', 'ARG1', 'x1')
     ]
     assert g.edges(target='e3') == [
-        ('e2', 'ARG2', 'e3'),
-        ('_sleep_v_1', 'instance', 'e3')
+        ('e2', 'ARG2', 'e3')
     ]